requirements.txt
pyyaml
pyrandom
uuid